index_selection/requirements.txt
# Core dependencies
pyyaml>=6.0
openai>=1.0.0

# Database
psycopg2-binary>=2.9.0

# Scientific (for analysis)
numpy>=1.24.0
